backend/requirements.txt
django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.3
reportlab>=4.0
Pillow>=10.0
python-dotenv>=1.0
gunicorn>=21.2
